verdere dependencies voor SpringBoot laagjesmodel toegevoegd. tevens bestaande televisionclass hernoemd en een nieuwe television model klasse aangemaakt
