+++ src/main/resources/application.properties
@@ -1,1 +1,15 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/tech_it_easy
+spring.datasource.username=postgres
+spring.datasource.password=wachtwoord
 
+#forces to generate the data structures
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=create
+
+#spring
+spring.sql.init.platform=postgres
+
+#JPA
+spring.jpa.database=postgresql
+#laat de sql commands zien
+spring.jpa.show-sql=true
